Rename some parameters, methods and package related to App Runtime Secrets
