+++ src/itest/java/com/iexec/sms/secret/ApplicationRuntimeSecretIntegrationTests.java
@@ -20,8 +20,8 @@ import com.iexec.common.utils.HashUtils;
 import com.iexec.sms.ApiClient;
 import com.iexec.sms.CommonTestSetup;
 import com.iexec.sms.encryption.EncryptionService;
-import com.iexec.sms.secret.applicationruntime.ApplicationRuntimeSecret;
-import com.iexec.sms.secret.applicationruntime.ApplicationRuntimeSecretRepository;
+import com.iexec.sms.secret.app.ApplicationRuntimeSecret;
+import com.iexec.sms.secret.app.ApplicationRuntimeSecretRepository;
 import feign.FeignException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
