+++ src/main/java/com/iexec/sms/secret/app/ApplicationRuntimeSecret.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.iexec.sms.secret.applicationruntime;
+package com.iexec.sms.secret.app;
 
 import com.iexec.sms.secret.Secret;
 import lombok.Data;
